Updated authentication with HTTPS cookie
Implemented better security for authentication handling. Now the token that is saved after a successful login is not stored in localStorage. Instead i added a HTTPS cookie for jwt (json web token) which functions as a refresh token.
Now there are 2 tokens. One refresh token and one access token. The access token is shortlived and is refreshed by the refresh token.